+++ src/testcurl/test_quiesce.c
@@ -347,7 +347,7 @@ ServeOneRequest (void *param)
       if ((WSAEINVAL != WSAGetLastError ()) ||
           (0 != rs.fd_count) || (0 != ws.fd_count) || (0 != es.fd_count) )
         externalErrorExitDesc ("Unexpected select() error");
-      Sleep (tv.tv_sec * 1000 + tv.tv_usec / 1000);
+      Sleep ((DWORD) (tv.tv_sec * 1000 + tv.tv_usec / 1000));
 #endif
     }
     MHD_run (d);
@@ -614,7 +614,7 @@ testExternalGet (void)
         if ((WSAEINVAL != WSAGetLastError ()) ||
             (0 != rs.fd_count) || (0 != ws.fd_count) || (0 != es.fd_count) )
           externalErrorExitDesc ("Unexpected select() error");
-        Sleep (tv.tv_sec * 1000 + tv.tv_usec / 1000);
+        Sleep ((DWORD) (tv.tv_sec * 1000 + tv.tv_usec / 1000));
 #endif
       }
       curl_multi_perform (multi, &running);
